+++ src/main/java/src/main/java/GUI/ImagePanel.java
@@ -1,4 +1,4 @@
-package src.main.java;
+package src.main.java.GUI;
 
 import javax.swing.JPanel;
